MedAI_Project (1)/MedAI/backend/requirements.txt
Flask
flask-cors
torch
torchvision
pydicom
Pillow
numpy
